refactor + do the things for all commands
