+++ latch_cli/exceptions/cache.py
@@ -1,0 +1,4 @@
+from typing import Dict, List, Tuple
+
+
+_code_cache: Dict[Tuple[str, int], List[str]] = {}
