+++ setup.py
@@ -19,7 +19,7 @@ setup(
     description="The Latchbio SDK",
     packages=find_packages(),
     include_package_data=True,
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     entry_points={
         "console_scripts": [
             "latch=latch_cli.main:main",
